refactor(api): use DEFAULT PASSWORD constant from common file
